Optimize queries, related to #24.
