+++ LobitaBot/LobitaBot/CommandHandler.cs
@@ -1,221 +1,12 @@
-﻿using Discord;
-using Discord.Commands;
+﻿using Discord.Commands;
 using Discord.WebSocket;
 using Microsoft.Extensions.DependencyInjection;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 
 namespace LobitaBot
 {
-    public class VideoService
-    {
-        public int RollIndex { get; set; } = -1;
-    }
-
-    public class PageData
-    {
-        public List<List<TagData>> Pages { get; set; }
-        public int PageNum { get; set; }
-        public DateTime DateTime { get; }
-        public bool AlphabeticallySorted { get; set; } = false;
-        public bool NumericallySorted { get; set; } = false;
-        public bool SortedAscending { get; set; } = false;
-
-        public PageData(List<List<TagData>> pages)
-        {
-            Pages = pages;
-            PageNum = 0;
-            DateTime = DateTime.Now;
-        }
-    }
-
-    public class PageService
-    {
-        public bool HandlerAdded { get; set; } = false;
-        public Dictionary<ulong, PageData> PageIndex { get; } = new Dictionary<ulong, PageData>();
-
-        public void AddLimited(ulong msgId, PageData pageData)
-        {
-            if (PageIndex.Count == 100)
-            {
-                ulong oldestMsg = PageIndex.Aggregate((x, y) => x.Value.DateTime < y.Value.DateTime ? x : y).Key;
-
-                PageIndex.Remove(oldestMsg);
-            }
-
-            PageIndex.Add(msgId, pageData);
-        }
-
-        public void SortAlphabeticalAsc(ulong msgId)
-        {
-            PageData pageData = PageIndex[msgId];
-            List<TagData> tagData = new List<TagData>();
-
-            foreach (List<TagData> page in pageData.Pages)
-            {
-                foreach (TagData t in page)
-                {
-                    tagData.Add(t);
-                }
-            }
-
-            tagData.Sort((t1, t2) => t1.TagName.CompareTo(t2.TagName));
-
-            PageIndex[msgId].Pages = TagParser.CompileSuggestions(tagData, EmbedBuilder.MaxFieldCount);
-            PageIndex[msgId].AlphabeticallySorted = true;
-            PageIndex[msgId].NumericallySorted = false;
-            PageIndex[msgId].SortedAscending = true;
-        }
-
-        public void SortAlphabeticalDesc(ulong msgId)
-        {
-            PageData pageData = PageIndex[msgId];
-            List<TagData> tagData = new List<TagData>();
-
-            foreach (List<TagData> page in pageData.Pages)
-            {
-                foreach (TagData t in page)
-                {
-                    tagData.Add(t);
-                }
-            }
-
-            tagData.Sort((t1, t2) => t2.TagName.CompareTo(t1.TagName));
-
-            PageIndex[msgId].Pages = TagParser.CompileSuggestions(tagData, EmbedBuilder.MaxFieldCount);
-            PageIndex[msgId].AlphabeticallySorted = true;
-            PageIndex[msgId].NumericallySorted = false;
-            PageIndex[msgId].SortedAscending = false;
-        }
-
-        public void SortPostNumAsc(ulong msgId)
-        {
-            PageData pageData = PageIndex[msgId];
-            List<TagData> tagData = new List<TagData>();
-
-            foreach (List<TagData> page in pageData.Pages)
-            {
-                foreach (TagData t in page)
-                {
-                    tagData.Add(t);
-                }
-            }
-
-            tagData.Sort((t1, t2) => t1.NumLinks.CompareTo(t2.NumLinks));
-
-            PageIndex[msgId].Pages = TagParser.CompileSuggestions(tagData, EmbedBuilder.MaxFieldCount);
-            PageIndex[msgId].AlphabeticallySorted = false;
-            PageIndex[msgId].NumericallySorted = true;
-            PageIndex[msgId].SortedAscending = true;
-        }
-
-        public void SortPostNumDesc(ulong msgId)
-        {
-            PageData pageData = PageIndex[msgId];
-            List<TagData> tagData = new List<TagData>();
-
-            foreach (List<TagData> page in pageData.Pages)
-            {
-                foreach (TagData t in page)
-                {
-                    tagData.Add(t);
-                }
-            }
-
-            tagData.Sort((t1, t2) => t2.NumLinks.CompareTo(t1.NumLinks));
-
-            PageIndex[msgId].Pages = TagParser.CompileSuggestions(tagData, EmbedBuilder.MaxFieldCount);
-            PageIndex[msgId].AlphabeticallySorted = false;
-            PageIndex[msgId].NumericallySorted = true;
-            PageIndex[msgId].SortedAscending = false;
-        }
-    }
-
-    public class CacheService
-    {
-        private List<PostData> tagCache = new List<PostData>();
-
-        public void Add(PostData pd)
-        {
-            tagCache.Add(pd);
-        }
-
-        public void Clear()
-        {
-            tagCache.Clear();
-        }
-
-        public bool CharacterInCache(string tagName)
-        {
-            return tagCache.Any(pd => pd.TagName == tagName);
-        }
-
-        public bool SeriesInCache(string seriesName)
-        {
-            return tagCache.Any(pd => pd.SeriesName == seriesName);
-        }
-
-        public bool CharacterAloneInCache(string tagName)
-        {
-            return !tagCache.Any(pd => pd.TagName != tagName);
-        }
-
-        public PostData CacheRandom()
-        {
-            PostData pd = null;
-
-            if (tagCache.Count > 0)
-            {
-                Random rand = new Random();
-                int index = rand.Next(0, tagCache.Count);
-
-                pd = tagCache[index];
-            }
-
-            return pd;
-        }
-
-        public PostData CacheNext(int index)
-        {
-            PostData pd = null;
-
-            if (index < tagCache.Count - 1 && index >= 0)
-            {
-                pd = tagCache[index + 1];
-            }
-            else if (index >= tagCache.Count - 1)
-            {
-                pd = tagCache.Last();
-            }
-
-            return pd;
-        }
-
-        public PostData CachePrevious(int index)
-        {
-            PostData pd = null;
-
-            if (index > 0 && index < tagCache.Count)
-            {
-                pd = tagCache[index - 1];
-            }
-            else if (index <= 0)
-            {
-                pd = tagCache.First();
-            }
-
-            return pd;
-        }
-
-        public int CacheSize()
-        {
-            return tagCache.Count();
-        }
-    }
-
     public class CommandHandler
     {
         private readonly CommandService commands;
