+++ LobitaBot/LobitaBot/Data/TagData.cs
@@ -1,0 +1,16 @@
+﻿namespace LobitaBot
+{
+    public class TagData
+    {
+        public TagData(string tagName, int tagID, long numLinks)
+        {
+            TagName = tagName;
+            TagID = tagID;
+            NumLinks = numLinks;
+        }
+
+        public string TagName { get; }
+        public int TagID { get; }
+        public long NumLinks { get; }
+    }
+}
