+++ LobitaBot/LobitaBot/Persistence/DbCharacterIndex.cs
@@ -22,7 +22,7 @@ namespace LobitaBot
                 $"FROM links AS l, tag_links AS tl, tags AS t, series_tags AS st, series AS s " +
                 $"WHERE l.id = tl.link_id AND t.id = tl.tag_id AND t.id = st.tag_id AND s.id = st.series_id AND t.name = '{searchTerm}'";
 
-            PopulateCache(postQuery);
+            PopulateCacheAsync(postQuery);
 
             return _cacheService.CacheRandom();
         }
@@ -40,7 +40,7 @@ namespace LobitaBot
                 $"FROM links AS l, tag_links AS tl, tags AS t, series_tags AS st, series AS s " +
                 $"WHERE l.id = tl.link_id AND t.id = tl.tag_id AND t.id = st.tag_id AND s.id = st.series_id AND t.name = '{searchTerm}'";
 
-            PopulateCache(postQuery);
+            PopulateCacheAsync(postQuery);
 
             return _cacheService.CacheNext(index);
         }
@@ -58,7 +58,7 @@ namespace LobitaBot
                 $"FROM links AS l, tag_links AS tl, tags AS t, series_tags AS st, series AS s " +
                 $"WHERE l.id = tl.link_id AND t.id = tl.tag_id AND t.id = st.tag_id AND s.id = st.series_id AND t.name = '{searchTerm}'";
 
-            PopulateCache(postQuery);
+            PopulateCacheAsync(postQuery);
 
             return _cacheService.CachePrevious(index);
         }
