+++ LobitaBot/LobitaBot/Services/VideoService.cs
@@ -1,0 +1,7 @@
+﻿namespace LobitaBot
+{
+    public class VideoService
+    {
+        public int RollIndex { get; set; } = -1;
+    }
+}
